Solved 3C with performance issues for the figure 8
