+++ src/tester1/annotation.go
@@ -1,15 +1,16 @@
 package tester
 
 import (
-	"sync"
+	"fmt"
 	"os"
 	"os/signal"
-	"fmt"
-	"time"
-	"strings"
 	"slices"
+	"strings"
+	"sync"
+	"time"
+
+	models "6.5840/models1"
 	"github.com/anishathalye/porcupine"
-	"6.5840/models1"
 )
 
 ///
@@ -76,10 +77,10 @@ func FinalizeAnnotations(end string) []porcupine.Annotation {
 
 	t := timestamp()
 	aend := porcupine.Annotation{
-		Tag: TAG_INFO,
-		Start: t,
-		Description: end,
-		Details: end,
+		Tag:             TAG_INFO,
+		Start:           t,
+		Description:     end,
+		Details:         end,
 		BackgroundColor: COLOR_INFO,
 	}
 	annotations = append(annotations, aend)
@@ -148,7 +149,7 @@ func AnnotateCheckerBegin(details string) {
 	defer finfo.mu.Unlock()
 
 	finfo.ckbegin = CheckerBegin{
-		ts: timestamp(),
+		ts:      timestamp(),
 		details: details,
 	}
 }
@@ -238,7 +239,7 @@ func AnnotateConnection(connection []bool) {
 
 func annotateFault() {
 	trues := make([]bool, finfo.nservers)
-	for id := range(trues) {
+	for id := range trues {
 		trues[id] = true
 	}
 	falses := make([]bool, finfo.nservers)
@@ -254,7 +255,7 @@ func annotateFault() {
 	crashes := make([]int, 0)
 	var builder strings.Builder
 	builder.WriteString("partition = ")
-	for id, connected := range(finfo.connected) {
+	for id, connected := range finfo.connected {
 		if finfo.crashed[id] {
 			crashes = append(crashes, id)
 			continue
@@ -289,11 +290,11 @@ func AnnotateClearFailure() {
 	finfo.mu.Lock()
 	defer finfo.mu.Unlock()
 
-	for id := range(finfo.crashed) {
+	for id := range finfo.crashed {
 		finfo.crashed[id] = false
 	}
 
-	for id := range(finfo.connected) {
+	for id := range finfo.connected {
 		finfo.connected[id] = true
 	}
 
@@ -305,7 +306,7 @@ func AnnotateShutdown(servers []int) {
 	defer finfo.mu.Unlock()
 
 	changed := false
-	for _, id := range(servers) {
+	for _, id := range servers {
 		if !finfo.crashed[id] {
 			changed = true
 		}
@@ -326,7 +327,7 @@ func AnnotateShutdownAll() {
 	finfo.mu.Unlock()
 
 	servers := make([]int, n)
-	for i := range(servers) {
+	for i := range servers {
 		servers[i] = i
 	}
 	AnnotateShutdown(servers)
@@ -337,7 +338,7 @@ func AnnotateRestart(servers []int) {
 	defer finfo.mu.Unlock()
 
 	changed := false
-	for _, id := range(servers) {
+	for _, id := range servers {
 		if finfo.crashed[id] {
 			changed = true
 		}
@@ -358,7 +359,7 @@ func AnnotateRestartAll() {
 	finfo.mu.Unlock()
 
 	servers := make([]int, n)
-	for i := range(servers) {
+	for i := range servers {
 		servers[i] = i
 	}
 	AnnotateRestart(servers)
@@ -379,13 +380,13 @@ func (an *Annotation) finalize() []porcupine.Annotation {
 	x := an.annotations
 
 	t := timestamp()
-	for tag, cont := range(an.continuous) {
+	for tag, cont := range an.continuous {
 		a := porcupine.Annotation{
-			Tag: tag,
-			Start: cont.start,
-			End: t,
-			Description: cont.desp,
-			Details: cont.details,
+			Tag:             tag,
+			Start:           cont.start,
+			End:             t,
+			Description:     cont.desp,
+			Details:         cont.details,
 			BackgroundColor: cont.bgcolor,
 		}
 		x = append(x, a)
@@ -409,10 +410,10 @@ func (an *Annotation) annotatePointColor(
 	an.mu.Lock()
 	t := timestamp()
 	a := porcupine.Annotation{
-		Tag: tag,
-		Start: t,
-		Description: desp,
-		Details: details,
+		Tag:             tag,
+		Start:           t,
+		Description:     desp,
+		Details:         details,
 		BackgroundColor: bgcolor,
 	}
 	an.annotations = append(an.annotations, a)
@@ -424,11 +425,11 @@ func (an *Annotation) annotateIntervalColor(
 ) {
 	an.mu.Lock()
 	a := porcupine.Annotation{
-		Tag: tag,
-		Start: start,
-		End: timestamp(),
-		Description: desp,
-		Details: details,
+		Tag:             tag,
+		Start:           start,
+		End:             timestamp(),
+		Description:     desp,
+		Details:         details,
 		BackgroundColor: bgcolor,
 	}
 	an.annotations = append(an.annotations, a)
@@ -446,8 +447,8 @@ func (an *Annotation) annotateContinuousColor(
 		// The first continuous annotation for tag. Simply add it to the
 		// continuous map.
 		an.continuous[tag] = Continuous{
-			start: timestamp(),
-			desp: desp,
+			start:   timestamp(),
+			desp:    desp,
 			details: details,
 			bgcolor: bgcolor,
 		}
@@ -458,17 +459,17 @@ func (an *Annotation) annotateContinuousColor(
 	// annotation and add this one to the continuous map.
 	t := timestamp()
 	aprev := porcupine.Annotation{
-		Tag: tag,
-		Start: cont.start,
-		End: t,
-		Description: cont.desp,
-		Details: cont.details,
+		Tag:             tag,
+		Start:           cont.start,
+		End:             t,
+		Description:     cont.desp,
+		Details:         cont.details,
 		BackgroundColor: cont.bgcolor,
 	}
 	an.annotations = append(an.annotations, aprev)
 	an.continuous[tag] = Continuous{
-		start: t,
-		desp: desp,
+		start:   t,
+		desp:    desp,
 		details: details,
 		bgcolor: bgcolor,
 	}
@@ -487,11 +488,11 @@ func (an *Annotation) annotateContinuousEnd(tag string) {
 	// End the on-going continuous annotation for tag.
 	t := timestamp()
 	aprev := porcupine.Annotation{
-		Tag: tag,
-		Start: cont.start,
-		End: t,
-		Description: cont.desp,
-		Details: cont.details,
+		Tag:             tag,
+		Start:           cont.start,
+		End:             t,
+		Description:     cont.desp,
+		Details:         cont.details,
 		BackgroundColor: cont.bgcolor,
 	}
 	an.annotations = append(an.annotations, aprev)
@@ -517,10 +518,10 @@ func (an *Annotation) cleanup(failed bool, end string) {
 
 	t := timestamp()
 	aend := porcupine.Annotation{
-		Tag: TAG_INFO,
-		Start: t,
-		Description: end,
-		Details: end,
+		Tag:             TAG_INFO,
+		Start:           t,
+		Description:     end,
+		Details:         end,
 		BackgroundColor: COLOR_INFO,
 	}
 	annotations = append(annotations, aend)
@@ -532,7 +533,7 @@ func (an *Annotation) cleanup(failed bool, end string) {
 		// Save the vis file in a temporary file.
 		file, err = os.CreateTemp("", "porcupine-*.html")
 	} else {
-		file, err = os.OpenFile(fpath, os.O_RDWR | os.O_CREATE | os.O_TRUNC, 0644)
+		file, err = os.OpenFile(fpath, os.O_RDWR|os.O_CREATE|os.O_TRUNC, 0644)
 	}
 	if err != nil {
 		fmt.Printf("info: failed to open visualization file %s (%v)\n", fpath, err)
@@ -549,10 +550,10 @@ func (an *Annotation) cleanup(failed bool, end string) {
 
 func mkAnnotation() *Annotation {
 	an := Annotation{
-		mu: new(sync.Mutex),
+		mu:          new(sync.Mutex),
 		annotations: make([]porcupine.Annotation, 0),
-		continuous: make(map[string]Continuous),
-		finalized: false,
+		continuous:  make(map[string]Continuous),
+		finalized:   false,
 	}
 
 	return &an
@@ -560,15 +561,15 @@ func mkAnnotation() *Annotation {
 
 func mkFrameworkInfo(nservers int) *FrameworkInfo {
 	conn := make([]bool, nservers)
-	for id := range(conn) {
+	for id := range conn {
 		conn[id] = true
 	}
 
 	finfo := FrameworkInfo{
-		mu: new(sync.Mutex),
-		nservers: nservers,
+		mu:        new(sync.Mutex),
+		nservers:  nservers,
 		connected: conn,
-		crashed: make([]bool, nservers),
+		crashed:   make([]bool, nservers),
 	}
 
 	return &finfo
@@ -578,7 +579,7 @@ func captureSignal() struct{} {
 	// Capture SIGINT to visualize on interruption.
 	c := make(chan os.Signal, 1)
 	signal.Notify(c, os.Interrupt)
-	go func(){
+	go func() {
 		for range c {
 			annotation.cleanup(true, "interrupted")
 			os.Exit(1)
